requirements.txt
gradio~=4.41.0
replicate~=0.31.0
requests~=2.32.3
pillow~=10.4.0